Show dwz stderr on failure
We redirect dwz stderr to a dwz.err file in case we're expecting some error
message, or want to grep for it, but in case the test fails in some way, we
still want to see dwz.err in the testrun output.

Fix this by adding 'cat dwz.err' where appropriate.

2019-03-16  Tom de Vries  <tdevries@suse.de>

	* testsuite/dwz.tests/objcopy-remove-debug-abbrev.sh: Emit dwz.err on
	exit.
	* testsuite/dwz.tests/objcopy-strip-debug.sh: Same.
	* testsuite/dwz.tests/pr24174.sh: Same.
	* testsuite/dwz.tests/too-many-dies.sh: Redirect stderr to dwz.err and
	grep for expected error message.  Emit dwz.err on exit.