Compile dwz-for-tests with -U__GNUC__
There are ifdefs in dwz.c testing for __GNUC__, which is a predefined gcc (and
clang) macro.  When compiling with make CC=gcc CFLAGS=-U__GNUC__, to trigger
the !__GNUC__ path, we run into some warnings.

Compile dwz-for-test with -U__GNUC__ to trigger the !__GNUC__ path in a regular
gcc/clang build-and test cycle, and fix the resulting warnings.

2019-03-29  Tom de Vries  <tdevries@suse.de>

	* Makefile (dwz-for-test): Marking warning flags identical to dwz
	target.  Compile with -U__GNUC__.
	* dwz.c [!__GNUC__] (likely, unlikely): Fix missing parentheses warning.
	(adjust_loclist): Fix unused parameter warning using void cast instead
	of __attribute__((unused)).