Add --devel-ignore-size
Dwz is a dwarf compression tool with the intent to transform dwarf into
equivalent but smaller dwarf.  Consequently:
- during transformation it takes decisions based on size heuristics, and
- at the end of the transformation it discards the result if it's not smaller
  than the original.

Currently, when we run the gdb testsuite in conjunction with dwz using
CC_WITH_TWEAKS_FLAGS, we only test dwz-transformed dwarf if in fact dwz
does not discard the result, and we only test the result of the
transformations enabled by the size heuristics.

Add a developer-only option --devel-ignore-size, that ignores size metrics
and heuristics.  With this option, we force dwz to:
- move all duplicate DIEs to partial units,
- factor out all common partial unit import sets into import-only partial
  units, and
- write out the result.

Also, this may prove useful in dwz test-cases to force a transformation if we
want to test that transformation.

2019-04-06  Tom de Vries  <tdevries@suse.de>

	* dwz.c (partition_dups_1, create_import_tree): If ignore_size, ignore
	size heuristic.
	(dwz): If ignore_size, skip size test.
	(dwz_options): Add --devel-ignore-size.
	* Makefile (TEST_EXECS): Add min.
	(min): New target.
	* testsuite/dwz.tests/devel-ignore-size.sh: New test.
	* testsuite/dwz.tests/min-2.c: New test source.
	* testsuite/dwz.tests/min.c: New test source.