Fix multifile opt if single file opt fails
When optimizing hello in multi-file mode we have:
...
$ cp hello 1
$ cp 1 2
$ dwz -m 3 1 2
$ ls 3
3
$
...

In multifile mode, dwz first optimizes all inputs in single file mode, and then
in multifile mode.

But if we try to factor out the single file optimization using a seperate dwz
invocation we get:
...
$ cp hello 1
$ dwz 1
$ cp 1 2
$ ./dwz -m 3 1 2
./dwz: 1: DWARF compression not beneficial - old size 436 new size 436
./dwz: 2: DWARF compression not beneficial - old size 436 new size 436
ls: cannot access '3': No such file or directory
$
...

Multifile optimization fails here, because optimize_multifile early-returns
-1 because multi_ehdr.e_ident[0] == '\0'.  The multi_ehdr variable is
initialized in write_dso in single file mode, which is not triggered because
compression is not beneficial.

Fix this by moving the multi_ehdr initialization from write_dso to
write_multifile.

2019-04-23  Tom de Vries  <tdevries@suse.de>

	PR dwz/24354
	PR dwz/24477
	* dwz.c (write_dso): Move multi_ehdr init ...
	(write_multifile): ... here.
	* testsuite/dwz.tests/twice-multifile.sh: Expect multifile optimization
	to work.