[low-mem] Fix DW_OP_GNU_implicit_pointer handling
When twice dwz-ing a cc1 build with bootstrap-lto in low-mem mode, we run
into this error:
...
$ dwz -l0 cc1
$ dwz -l0 cc1
dwz: cc1: Couldn't find DIE referenced by DW_OP_GNU_implicit_pointer
...

The error is triggered as follows.  There's a DIE 0x369186:
...
 <12><369186>: Abbrev Number: 3 (DW_TAG_formal_parameter)
     <369187>   DW_AT_abstract_origin: <0xecc97>
     <36918b>   DW_AT_location    : 0x4ee259 (location list)
     <36918f>   DW_AT_GNU_locviews: 0x4ee255
...
referring to location list 0x4ee259:
...
    004ee255 v000000000000004 v000000000000001 location view pair
    004ee257 v000000000000000 v000000000000000 location view pair

    004ee259 v000000000000004 v000000000000001 views at 004ee255 for:
             0000000000876020 000000000087602b (DW_OP_GNU_implicit_pointer: \
                                                <0x800> 0)
    004ee271 v000000000000000 v000000000000000 views at 004ee257 for:
             0000000000587ca8 0000000000587cb0 (DW_OP_GNU_implicit_pointer: \
                                                <0x800> 0)
    004ee289 <End of list>
...
which refer to DIE 0x800 in a partial unit 0x7f7:
...
 <0><7f7>: Abbrev Number: 5 (DW_TAG_partial_unit)
    <7f8>   DW_AT_stmt_list   : 0xd1db8
    <7fc>   DW_AT_comp_dir    : (indirect string, offset: 0x2183): \
                                /home/mliska/Programming/gcc/objdir/gcc
 <1><800>: Abbrev Number: 112 (DW_TAG_dwarf_procedure)
    <801>   DW_AT_location    : 21 byte block: 9e 13 66 69 72 73 74 5f 69 6d \
                                6d 5f 75 73 65 5f 73 74 6d 74 0 \
                                (DW_OP_implicit_value 19 byte block: 66 69 72 \
                                73 74 5f 69 6d 6d 5f 75 73 65 5f 73 74 6d 74 0 )
...

When calculating the checksum for DIE 0x369186 in checksum_die, we call
read_loclist for loc_list 0x4ee259 and read_exprloc for
DW_OP_GNU_implicit_pointer 0x800 0, after which we run into the error for
addr == 0x800 here:
...
	  ref = off_htab_lookup (NULL, addr);
	  if (ref == NULL)
	    {
	      error (0, 0, "%s: Couldn't find DIE referenced by %s",
		     dso->filename, get_DW_OP_str (op));
	      return 1;
	    }
...

In read_debug_info in regular mode, the .debug_info section is processed one
CU at a time.  Each CU is parsed, after which the checksums and duplicates are
calculated, but the parsed CU remains in memory when processing the next CU.
So, when parsing the partial unit containing the DIE at 0x800, the DIE's added
to off_htab, and when reading the CU containing DIE 0x369186, calculating the
checksum for DIE 0x369186 in checksum_die and trying to find the DIE at 0x800
in read_exprloc, we succeed because DIE 0x800 is still present in off_htab.

In read_debug_info in low-mem mode, also the .debug_info section is processed
one CU at a time. Each CU is parsed, after which the checksums and duplicates
are calculated, but after that the amount of parsed CU that remains in memory
is minimized by remove_unneeded and collapse_children.  The inter-cu
referenced DIEs however need to be kept in memory, because they may be
referenced from CUs yet to process.  But there's no way of knowing which DIEs
are inter-cu referenced, before all CUs have been parsed.  To solve this
problem, in low-mem mode read_debug_info is split up into two phases which
each read the .debug_info section once.  In phase 1, for each DIE that is
inter-cu referenced, a dummy die is allocated, which is merged during phase 2
with the actual DIE when it's being parsed.  This way in phase 2 we can always
know if a DIE is inter-cu referenced, and consequently keep it in memory.

However, in phase 1 we currently only handle inter-cu refs of one type:
DW_FORM_ref_addr refs from DIE attributes, and consequently no dummy DIE is
generated during phase 1 for DIE 0x800, so it's not kept during phase 2, which
leads to the error.

Fix this by also creating dummy DIEs in phase 1 for inter-cu references in
expr_locs.  [ And due to the fact that we're not keeping track of the CUs of
dummy DIEs, we cannot detect when references which could be inter-cu, are in
fact intra-cu, so we also add dummy DIEs for those. ]

2019-04-29  Tom de Vries  <tdevries@suse.de>

	PR dwz/24204
	* dwz.c (read_exprloc_low_mem_phase1, read_loclist_low_mem_phase1)
	(add_locexpr_dummy_dies): New function.
	(read_debug_info): Call add_locexpr_dummy_dies.