Make updating sections after section header string table order-independent
Function write_dso updates the section header table: it adds, removes and
updates sections.  It does this in two phases:
- it calculates the effect of changing sections on the following
  sections sh_offset (as well as on the section header table e_shoff)
- it makes sure sh_offset is sh_addralign aligned (as well as that the section
  header table e_shoff is ELFCLASS-appropriately aligned).

The first phase is independent of the order of the section header table, apart
from this loop:
...
  for (j = dso->ehdr.e_shstrndx + 1; j < dso->ehdr.e_shnum; ++j)
    dso->shdr[j].sh_offset += len;
...
It adds an increase of the size of the section header string table to
sections 'after' the section header string table.  But the after test here is
implemented in terms of order in the section header table, which only works if
if the section header table is sorted.

Make the first phase order-independent by rewriting the after test in terms of
sh_offset.

2019-06-25  Tom de Vries  <tdevries@suse.de>

	* dwz.c (write_dso): Make updating of sections after section header
	string table robust against unsorted section header table.