Require x86_64 for pr24172.sh
When running the tests on openSUSE Factory ppc64, I run into a FAIL for
pr24172.sh.  The problem is that we expect an error, but dwz succeeds
instead.  The .S test-case contains an incorrect encoding, so we can
expect different behaviour on different systems (pointer size, endiannes).

Fix the FAIL by requiring x86_64 for pr24172.sh.

2019-06-27  Tom de Vries  <tdevries@suse.de>

	* testsuite/dwz.tests/dwz-tests.exp: Require x86_64 for pr24172.sh.