Fix dynamic-stack-buffer-overflow (write_dso, distance)
When compiling with gcc 8 and -fsanitize=address -lasan, we run into a
dynamic-stack-buffer-overflow related to the distance variable in write_dso.

The variable is allocated with dimension dso->ehdr.e_shnum, but we use
dso->ehdr.e_shnum elements while not using element 0.

Fix this by increasing the size of the distance variable.

This fixes the dwz-fedora-s390x buildbot failure.

2019-06-27  Tom de Vries  <tdevries@suse.de>

	PR dwz/24734
	* dwz.c (write_dso): Increase size of distance.