Handle running configure in source dir
The configure script was added to facilitate building in a separate build dir
rather than in the source dir.  It does this by generating a Makefile in the
build dir that calls make using the Makefile in the source dir.

However, running the configure in the source dir overwrites the Makefile in
the source dir.

Fix this by detecting running configure in the source dir, and not generating
the unnecessary Makefile.

2019-07-02  Tom de Vries  <tdevries@suse.de>

	* configure: Handle running in source dir.