Don't expect smaller executable when using objcopy and eu-unstrip
When testing using gold:
...
$ make CC="gcc -fuse-ld=gold" check
...
we run into:
...
FAIL: testsuite/dwz.tests/objcopy-eu-unstrip.sh
...

The problem is that the unstripped executable is larger than the original one.
This is a known problem, unrelated to dwz (filed as PR elfutils/24809).

Fix this by skipping the smaller-than.sh check when using objcopy and
eu-unstrip.

2019-07-15  Tom de Vries  <tdevries@suse.de>

	* testsuite/dwz.tests/objcopy-eu-unstrip-multifile.sh: Comment out
	smaller-than.sh check.
	* testsuite/dwz.tests/objcopy-eu-unstrip.sh: Same.