Add --devel-unoptimized-multifile
There is functionality enabled by -DDEBUG_OP_MULTIFILE.  This skips the
optimization step in optimize_multifile, and instead copies the aggregate file
(the virtual file described by the per-section temporary files) to the
multifile.

The aggregate file ends with a compilation unit without any DIE, so we run
into this error:
...
$ cp hello 1
$ cp 1 2
$ dwz -m 3 1 2
./dwz: 3: .debug_info section chunk doesn't contain a single compile_unit or \
  partial_unit
...
which has the effect that the finalize_multifile step is skipped as well.

The DWARF in the resulting multifile is the same as (using dwz without
-DDEBUG_OP_MULTIFILE) in dwz.debug_all:
...
$ cp hello 1
$ cp 1 2
$ dwz -m 3 1 2 --devel-save-temps
$ ../src/contrib/gen-dwz-debug-all.sh
...

The only difference is that:
- dwz.debug_all is created directly by objcopy from the per-section temporary
  files, and
- the multifile is created by dwz after reading in the per-section temporary
  files.

Given the almost duplicate functionality, I'm not sure if we still need this.
But for now, keep this functionality, and make it available using
--devel-unoptimized-multifile.

2019-11-06  Tom de Vries  <tdevries@suse.de>

	* dwz.c (unoptimized_multifile): New var.
	(optimize_multifile): Handle unoptimized_multifile.
	(dwz_options): Add --devel-unoptimized-multifile entry.