Clear struct dw_die's u.p2 field before use in 'phase 2'
In struct dw_die we have a union u with fields p1 and p2.  The p1 field is
used during 'phase 1', after which the space is reused for the p2 field during
'phase 2'.

While developing a patch for dwz, I ran into a case where a
'die->u.p2.die_new_offset != 0' assert in write_die did not trigger, because
die->u.p1 was used during 'phase 1' which left a non-zero value in
die->u.p2.die_new_offset.

Add initialization that clears the 'phase 2' field u.p2 before 'phase 2',
to increase the chance of triggering this type of assert.

Enable this only with -DDEVEL, since it increases runtime with ~1%.

2019-11-08  Tom de Vries  <tdevries@suse.de>

	* dwz.c (clear_p2_field): New function.
	(write_multifile, dw, optimize_multifile): Call clear_p2_field.