Add --devel-partition-dups-opt
There's an optimization condition here in partition_dups_1:
...
      /* If during second_phase there are some DIEs we want to force
         into a partial unit because they are referenced from something
         already forced into a partial unit, but also some DIEs with
         the same set of referrers, try to see if we can put also those
         into the partial unit.  They can be put there only if they
         don't refer to DIEs that won't be put into partial units.  */
      if (second_phase && force && force < j - k)
        {
...

However, due to a preceding loop:
...
      for (k = i; k < j; k++)
...
and a 'j > i' invariant at the start of that loop we'll always end up with
'j == k' at the condition, so the optimization will never be triggered.

The intention of the optimization condition is to test whether there are some
forced as well as some unforced DIEs in the partition [i,j).

The '&& force' part of the condition tests whether there are forced dies,
given that 'force' is the number of forced DIEs.

The '&& force < j - k' part of the condition is supposed to test whether there
are any unforced DIEs.  The correct way to test this is '&& force < j - i',
given that 'j - i' is the total number of DIEs in the partition.

However, with this fix in place, the optimization seems hard to trigger.  It
doesn't trigger in the regular testsuite, and while it triggers in the external
testsuite on vmlinux-4.20.13-0-vanilla.debug, the optimization has no
effect on the dwz output for that file.

In the absence of proof that the optimization is properly tested, classify the
optimization as experimental, and guard with --devel-partition-dups-opt.

2019-11-18  Tom de Vries  <tdevries@suse.de>

	PR dwz/24399
	* dwz.c (partition_dups_opt): New var.
	(partition_dups_1): Fix optimization condition.  Guard
	optimization with partition_dups_opt.
	(dwz_options): Add --devel-partition-dups-opt entry.