Fix ignored return value warning for link in write_dso
When compiled on ubuntu 18.04, we run into this warning:
...
dwz.c: In function ‘write_dso’:
dwz.c:11055:7: warning: ignoring return value of ‘link’, declared with \
  attribute warn_unused_result [-Wunused-result]
       link (dso->filename, buf);
...
in the --devel-save-temps code.

When fixing this warning by erroring out if the link call fails, we run into
the problem that the code containing the link call is called twice:
- once after regular mode, to save the resulting file, and
- once in multifile_finalize mode, which is unnecessary.

After fixing the code not to run in multifile_finalize mode, we run into the
problem that the link target may already exist, so remove it before calling
link.

2019-11-23  Tom de Vries  <tdevries@suse.de>

	* dwz.c (write_dso): Add and use save_to_temp parameter.  Unlink before
	linking.  Test result of link call.
	(dwz): Add arg to write_dso call.