Avoid unnecessary .debug_info section parsing
In regular mode, processing cc1 takes ~7s:
...
$ time.sh dwz cc1 -o 1 -lnone
maxmem:	1178008
real:	6.95
user:	6.19
sys:	0.75
...

OTOH, in low-mem mode, processing takes ~11.3s (~63% slower), but at a ~45%
lower memory usage:
...
$ time.sh dwz cc1 -o 1 -l0
maxmem:	655792
real:	11.37
user:	10.71
sys:	0.66
...

However, if we do not enforce either regular mode (through -lnone) or low-mem
mode (through -l0), we first try regular mode, until we run into the low-mem
limit, and then start over and continue in low-mem mode:
...
$ time.sh dwz cc1 -o 1
maxmem:	994704
real:	15.28
user:	14.31
sys:	0.96
...
This gives a longer execution time than enforced low-mem, and doesn't reduce
memory usage as much as enforced low-mem.  The default low-mem limit is 10
million, and the number of DIEs in cc1 is 10.2 million, so it takes a long
time in regular mode to find out that we're hitting the limit.

We might employ a strategy where we count the DIEs before entering either
low-mem or regular mode, but that itself again adds a penalty of ~5% in case
we don't run into the low-mem limit.

Use the estimated number of DIEs to determine whether we might run into the
low-mem limit, and if so, count the DIEs to make sure that that's the case,
and handle things accordingly.

This gives us an execution time and memory usage similar to enforced low-mem:
...
$ time.sh dwz cc1 -o 1
maxmem:	655256
real:	11.66
user:	11.01
sys:	0.64
...

Do the same for the max-die limit.

2019-11-23  Tom de Vries  <tdevries@suse.de>

	* dwz.c (try_debug_info): New function.
	(read_debug_info): If estimated number of DIEs in .debug_info exceeds
	either low_mem_die_limit or max_die_limit, use try_debug_info to
	verify this.