Only apply die limits in regular mode
In off_htab_add_die we have:
...
      if (low_mem
         || op_multifile
         || estimated_nr_dies >= low_mem_die_limit
         || estimated_nr_dies >= max_die_limit)
       initial_size = default_initial_size;
...

Die limits are only supposed to apply in regular mode, but the formulation
allows the limits to apply in read_multifile and finalize_multifile mode as
well.

I don't think this can trigger a difference in behaviour in finalize_multifile
mode, but it might in read_multifile mode.

Fix this by only testing die limits in regular mode.

2019-11-25  Tom de Vries  <tdevries@suse.de>

	* dwz.c (off_htab_add_die): Only test die limits in regular mode.