Only compile test-cases using hardcoded assembly for x86_64
Currently the test-cases using hardcoded assembly are tentatively compiled
for all architectures, and a failure to compile is handled as UNSUPPORTED.

Since the hardcoded assembly is generated on x86_64, this approach produces
a lot of warning on other architectures, which make it harder to spot actual
problems in the log.  Also, this approach might mask compilations error on
x86_64.

Fix this by only compiling these test-cases for x86_64, and expecting a valid
result for x86_64.

2019-12-12  Tom de Vries  <tdevries@suse.de>

	* Makefile (UNAME): New variable.
	(dw2-restrict, dw2-skip-prologue, py-section-script)
	(implptr-64bit-d2o4a8r8t0, varval): Only compile for x86_64.