Add field incoming_tail to struct import_cu
The struct import_cu has a field incoming, that describes a single linked list
of struct import_edge, sorted on ascending element.icu->idx.

Add a field incoming_tail to struct import_cu, to point to the tail of the
incoming list.  This allows access to the last element (with the highest idx)
without having to iterate the list from head to tail.

This patch has been additionally tested with --devel-verify-edges on by
default.

2019-12-17  Tom de Vries  <tdevries@suse.de>

	* dwz.c (struct import_cu): Add incoming_tail field.
	(verify_edges_1): Add verification of incoming_tail field.
	(remove_import_edges, create_import_tree): Handle incoming_tail field.