Fix bitvector_clear_bits
The commit 94e7891 "Reimplement seen flag of struct import_cu as bitvector"
introduces the function bitvector_clear_bits.

This function calculates the start element to be cleared, and the number of
elements, but then proceeds to use the number of elements as number of bytes
argument to the memset call.

Fix this by converting the memset argument from number of elements into number
of bytes.

This brings the improvement by the mentioned commit of 15% on vmlinux down
to 4%:
...
real:  mean:  16684.40 100.00%  stddev:  219.82
       mean:  15962.80  95.68%  stddev:  161.35
user:  mean:  16239.40 100.00%  stddev:   93.90
       mean:  15556.00  95.79%  stddev:  155.83
sys:   mean:    326.40 100.00%  stddev:   21.09
       mean:    313.60  96.08%  stddev:   33.66
...

2020-01-06  Tom de Vries  <tdevries@suse.de>

	* dwz.c (bitvector_clear_bits): Fix cleared range.