[odr] Combine decls duplicate chain with def duplicate chain
Add an optimization mode ODR_LINK in which we combine the decl chain with one
of the defs chains.

The duplicate chain containing the decls, combined with one type of def cannot
be handled by the backend if it starts with a decl.  So, we reorder it to
make sure it doesn't start with a decl.

We do this reordering ALAP, because the order of the duplicate chain is
required to be in increasing die_cu (die)->cu_chunk order starting from
the qsort in partition_dups (which sorts the array with duplicate chains into
equivalence classes with the same set of referrer CUs) up until the outer
loops of partition_dups_1 (which determine the equivalence class boundaries in
the array with duplicate chains).

2020-01-06  Tom de Vries  <tdevries@suse.de>

	* dwz.c (enum odr_mode): New enum.
	(merge_dups): New function.
	(split_dups): Call merge_dups if odr_mode != ODR_BASIC.
	(reorder_dups): New function.
	(partition_dups_1): Call reorder_dups if odr_mode != ODR_BASIC.