Cache file name property in struct dw_file
When profiling with benchmark clang-10, we observe a hotspot in the strlen
call in die_eq_1:
...
       │                    file_len = strlen (cu_file1->file);
       │        or     $0xffffffffffffffff,%rcx
  0,01 │        mov    %r15,%rdi
       │                    if (cu_file1->dir != NULL)
  0,01 │        mov    0x0(%r13),%r13
       │                    file_len = strlen (cu_file1->file);
 26,24 │        repnz  scas %es:(%rdi),%al
  3,16 │        mov    0x58(%rsp),%r10
...
That is, 29.4% of the time spent in die_eq_1 is spent in this strlen call
(while 23.10% of the overall time is spent in die_eq_1).

The file_len variable is used to calculate a more complex property of the
filename:
...
              else if (cu_file1->file[0] == '<'
                       && cu_file1->file[file_len - 1] == '>'
                       && strchr (cu_file1->file, '/') == NULL
...
which is also calculated in checksum_die.

Cache the property in a new struct dw_file field named
file_angle_brackets_encapsulated_no_slash, and used it in both checksum_die
and die_eq_1.

This speeds up processing of clang-10 with ~6.5%:
...
real:  mean:  263942.00  100.00%  stddev:  1558.78
       mean:  247611.33   93.81%  stddev:   687.17
user:  mean:  253755.33  100.00%  stddev:   447.36
       mean:  236928.33   93.37%  stddev:  1721.72
sys:   mean:    5296.00  100.00%  stddev:   820.97
       mean:    5396.67  101.90%  stddev:  1013.13
...

In particular, with --devel-progress we observe going from:
...
partition_dups split_dups
user: 94.50
sys : 0.16
...
to:
...
partition_dups split_dups
user: 74.44
sys : 0.01
...
which is a reduction of 21.2%.

2020-01-07  Tom de Vries  <tdevries@suse.de>

	* dwz.c (struct dw_file): Add field
	file_angle_brackets_encapsulated_no_slash.
	(read_debug_line): Initialize new field.
	(checksum_die, eq_1): Use new field.