Fix reference from PU to CU (--odr-mode=basic)
The function partition_dups is executed in two phases.  In phase 1 we decide
which DIEs we want to move into partial units because it looks profitable,
and in phase 2 we find which DIEs we have to move into partial units because
they are referenced by other DIEs already in partial units.

The process of splitting an odr duplicate chain (before phase 1) can result
in some of the DIEs not being part of any duplicate chain afterwards.  If
such a DIE is marked as referenced in phase 2, it won't be moved to a partial
unit, causing a reference from PU to CU.

Fix this by making sure that all the elements of an odr duplicate chain are
part of a duplicate chain after splitting, even if that is a chain of length
one (which we call a singleton).  Also make sure that all other singleton DIEs
referenced from singleton DIEs are added as duplicate chain.

The fix means that split_dups no longer returns NULL, which allows some
simplification in partition_dups, but we'll deal with that in a follow-up
patch, to keep this patch shorter and clearer.

2020-01-20  Tom de Vries  <tdevries@suse.de>

	PR dwz/25411
	* dwz.c (mark_singletons): New function.
	(split_dups): Add singleton DIE as duplicate chain.
	(reset_die_ref_seen): New function.
	(partition_dups): Call reset_die_ref_seen.  Set reset_die_ref to 1 for
	singleton DIEs.  Call mark_singletons for singleton DIEs.